python_app/requirements.txt
flask
kafka-python
#flask-cors
cassandra-driver
requests
